'Focus Event' on a D_Geometry object causes it parent gui component to gain focus
When a geometry is focused or under goes an event that causes the geometry to gain focus, the component that the geometry belongs to (accessed by getParent() on geometry) will get focused. i.e clicking on a geometry causes its parent to gain focus and not the geometry. Also the focus gain and loose events will not be invoked on the geometry but on its parent.